Move flox_instance
